Create Personas_LautaroCano.py
Carry out an algorithm that allows to register, cancel or modify a person's data.
The data of the person will be DNI, Last Name, Second Last Name, First Name, Middle Name and Third Name, Address, Contact Telephone Number and a list of the schools they attended.
The DNI cannot be modified. If the user makes a mistake when entering it, they must cancel it and then register it again. The Name, Surname and list of schools CAN be modified.
The Name and Surname must be required, not the second surname and the second and third name.